+++ web/src/store/requests/index.ts
@@ -27,7 +27,7 @@ export const parseResponse = async (response: Response, functionName: string) =>
       const errorMessage = json || {
         code: response.status,
         message: 'Unknown error occurred',
-        details: body
+        details: body,
       }
       const error = genericErrorMessageConstructor(functionName, errorMessage)
       throw new Error(error)