+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Changelog
 
-## [Unreleased](https://github.com/MarquezProject/marquez/compare/0.43.1...HEAD)
+## [Unreleased](https://github.com/MarquezProject/marquez/compare/0.45.0-rc.1...HEAD)
+
+## [0.45.0-rc.1](https://github.com/MarquezProject/marquez/compare/0.44.0...0.45.0-rc.1) - 2024-02-13
+
+### Added
+
+* Web: updates for Table and Column Lineage [`#2725`](https://github.com/MarquezProject/marquez/pull/2725) [@phixMe](https://github.com/phixMe)  
+    *Creates a new page for column lineage and an updated view for lineage with a common set of shared principles.*
+* Web: add dataset field level tags to UI [`#2729`](https://github.com/MarquezProject/marquez/pull/2729) [@davidsharp7](https://github.com/davidsharp7)  
+    *Updates the `DatasetTags` component to allow for field-level tagging/deletion and adds this to the `DatasetInfo` component.*
+
+### Fixed
+
+* Web: minor UI enhancements [`#2727`](https://github.com/MarquezProject/marquez/pull/2727) [@phixMe](https://github.com/phixMe)  
+    *Hygienic cleanup of project as a follow-up to [`#2725`](https://github.com/MarquezProject/marquez/pull/2725), including a fix for [`#2747`](https://github.com/MarquezProject/marquez/issues/2747).*
+
+## [0.44.0](https://github.com/MarquezProject/marquez/compare/0.43.1...0.44.0) - 2024-01-22
+
+### Added
+
+* Web: add dataset tags tabs for adding/deleting of tags [`#2714`](https://github.com/MarquezProject/marquez/pull/2714) [@davidsharp7](https://github.com/davidsharp7)  
+    *Adds a dataset tags component so that datasets can have tags added/deleted.*
+* API: Add endpoint to delete field-level tags [`#2705`](https://github.com/MarquezProject/marquez/pull/2705) [@davidsharp7](https://github.com/davidsharp7)  
+    *Adds delete endpoint to remove dataset field tags.*
+
+### Fixed
+
+* Web: fix dataset tag reducers bug [`#2716`](https://github.com/MarquezProject/marquez/pull/2716) [@davidsharp7](https://github.com/davidsharp7)  
+    *Removes result from dataset tags reducer to fix a sidebar bug.*
 
 ## [0.43.1](https://github.com/MarquezProject/marquez/compare/0.43.0...0.43.1) - 2023-12-20